refactor(cli): extract SetupSprite duplication into shared sprite.go
Create internal/cli/sprite.go with parameterized SetupSpriteWithConfig()
function that handles both start and resume Sprite setup modes. This
extracts ~80% duplicated code from start.go and resume.go.

Changes:
- Add SpriteSetupMode enum (Start/Resume) for mode-specific behavior
- Add SpriteSetupConfig struct with all setup parameters
- Move HandleServerPassword to sprite.go (shared between start/resume)
- Move checkoutBranch helper to sprite.go
- Update start.go, resume.go, review.go, update.go to use new shared code
- Update tests to use HandleServerPassword instead of handleResumeServerPassword

Net reduction of ~68 lines while consolidating duplicated logic.
